% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned.dist.matrix.R
\name{binned.dist.matrix}
\alias{binned.dist.matrix}
\title{binned.dist.matrix}
\usage{
binned.dist.matrix(x, group_column, target_columns, by = 0.1)
}
\arguments{
\item{x}{data frame with a grouping variable, and at least one numeric variable to be used for comparing distribution}

\item{group_column}{column name of x which defines the group}

\item{target_columns}{column name of x which will be used to compare distribution}

\item{by}{bin size to be used in binned.dist.matrix, Default: 0.1}
}
\value{
matrix(nrow=length(unique(x[[group_column]])),ncol=as.integer(length(1/by)) + 1)
}
\description{
historical binning method to obtain distance vector for different group value
}
\details{
matrix element value contains proportion of data contained in the bin region, and is scaled afterwards. NA and Inf values are eliminated.
}
\examples{
\dontrun{
if(interactive()){
 binned.dist.matrix(iris,'Species','Sepal.Length')
 }
}
}
\seealso{
\code{\link[dplyr]{tally_}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter_all}},\code{\link[dplyr]{all_vars}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}}
}
