\name{clusterd-package}
\alias{clusterd-package}
\alias{clusterd}
\docType{package}
\title{
\packageTitle{clusterd}
}
\description{
\packageDescription{clusterd}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{clusterd}
\packageIndices{clusterd}
 method to find out how similar is the distribution of attribute values among different groups : cluster.by.distribution
}
\author{
\packageAuthor{clusterd}

Maintainer: \packageMaintainer{clusterd}
}
\examples{
  cluster.by.distribution(iris,'Species','Sepal.Length')
}
