% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.clara.object.R
\name{get.clara.object}
\alias{get.clara.object}
\title{get.clara.object}
\usage{
get.clara.object(is.k.predefined, target.matrix, defined.k = numeric(0),
  clara.metric = "euclidean", clusgap.B = 400)
}
\arguments{
\item{is.k.predefined}{if TRUE, uses defined.k as the number of medoids for clustering}

\item{target.matrix}{distance matrix obtained from binned.dist.matrix(...)}

\item{defined.k}{if numeric(0) optimal number of medoids are calculated using cluster::clusGap and cluster::maxSE, Default: numeric(0)}

\item{clara.metric}{metric parameter given to cluster::clara Default: 'euclidean'}

\item{clusgap.B}{metric parameter given to cluster::clusGap, Default: 400}
}
\value{
clara
}
\description{
obtain clara object defined in cluster::clara using distance matrix obtained from binned.dist.matrix(...).
}
\details{
list(clara,silinfo,clustering)
}
\examples{
\dontrun{
if(interactive()){
 get.clara.object(F,binned.dist.matrix(iris,'Species','Sepal.Length'))
 }
}
}
\seealso{
\code{\link[cluster]{clusGap}},\code{\link[cluster]{clara}}
 \code{\link[dplyr]{reexports}},\code{\link[dplyr]{arrange}}
}
