% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_tibble.R
\docType{data}
\name{random.tibble}
\alias{random.tibble}
\title{random.tibble}
\format{A data frame with 52500 rows and 2 variables:
\describe{
  \item{\code{label}}{integer grouping label}
  \item{\code{rand}}{double value genrated from runif,rnorm, or rchisq}
}}
\usage{
random.tibble
}
\description{
ROWS GENERATED FROM UNIFORM,NORMAL,CHI-SQUARE DISTRIBUTION WITH DIFFERING PARAMETERS COMBINED AS SINGLE DATAFRAME.
}
\keyword{datasets}
